Update playbook handling
Keep frozen copies of the component's playbooks inside each Trial Network, in their separate folders
