+++ .gitignore
@@ -102,3 +102,4 @@ ENV/
 .flaskenv
 REPOSITORIES
 SETTINGS/*.yml
+TRIAL_NETWORKS
