+++ core/Api/playbook.py
@@ -4,7 +4,7 @@ sys.path.append('..')
 from uuid import UUID
 from flask import jsonify
 from flask_restx import Namespace, Resource, abort, reqparse
-from Library import Library
+from shared import Library
 
 api = Namespace('playbook', "Component's playbook and 6G-Library handling")
 
@@ -12,5 +12,4 @@ api = Namespace('playbook', "Component's playbook and 6G-Library handling")
 @api.route("/<string:text>")
 class TestbedResource(Resource):
     def get(self, text: str):
-        Library.UpdateLocalRepository()
         return jsonify({'text': text})
