--- core/Library/__init__.py
@@ -1,1 +1,0 @@
-from .library import Library
