+++ shared/__init__.py
@@ -2,6 +2,6 @@ from .log import Log, Level
 from .io import IO
 from .child import Child
 from .cli_executor import Cli
-from .library import Library, Playbook
+from .library import Library, Repository, Component
 from .rest_client import RestClient, Payload
 from .config_base import ConfigBase, Validable, RestApiConf, LoginRestApiConf, EnabledLoginRestApiConf
